pyproject.toml
[tool.poetry]
name = "strava-customgpt-action"
version = "0.1.0"
description = "A CustomGPT Action that connects to Strava and fetches information to be used for training follow-up and planning."
authors = ["Marco Trincavelli <trinchetto@gmail.com>"]
readme = "README.md"
packages = [
    { include = "strava_customgpt_action", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.12"
stravalib = "^2.4"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4"
ruff = "^0.4"
mypy = "^1.10"
pre-commit = "^3.7"
types-requests = "^2.32.0"

[tool.poetry.scripts]
strava-recent-activities = "strava_customgpt_action.cli:main"
strava-activities-api = "strava_customgpt_action.server:main"
strava-auth = "strava_customgpt_action.auth:main"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E203", "E501"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict_optional = true
warn_unused_configs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
